Change from middleware to proxy
